requirements.txt
attrs==25.4.0  
autobahn==25.10.2  
Automat==25.4.16  
certifi==2025.4.26  
cffi>=2.0.0  
charset-normalizer==3.4.2  
constantly==23.10.4
cryptography==46.0.3  
hyperlink==21.0.0  
idna==3.11
incremental==24.7.2  
isodate==0.7.2  
logzero==1.7.0  
numpy>=1.24.0
pandas>=2.0.0
pycparser==2.23  
pyotp==2.9.0  
pyparsing==3.2.5  
python-dateutil==2.9.0  
python-dotenv>=1.0.0
pytz==2025.2 
rdflib==7.4.0 
rdflib-jsonld==0.6.2 
requests==2.32.5  
simplejson==3.20.2  
six==1.17.0  
smartapi-python==1.5.5 
streamlit>=1.28.0
Twisted==25.5.0 
txaio==25.9.2  
typing-extensions==4.15.0  
urllib3==2.5.0  
websocket-client==1.9.0  
yfinance>=0.2.0
beautifulsoup4>=4.12.0
curl-cffi>=0.5.0
frozendict>=2.3.0
multitasking>=0.0.11
peewee>=3.17.0
platformdirs>=3.0.0
protobuf>=4.21.0
websockets>=10.0
zope.interface==8.1.1
pywebview>=5.2
